Add file for test helpers and DFT function
