+++ math/src/fft/fft_iterative.rs
@@ -94,8 +94,8 @@ where
 #[cfg(test)]
 mod tests {
     use crate::fft::helpers::log2;
+    use crate::fft::test_helpers::naive_matrix_dft_test;
     use crate::field::test_fields::u64_test_field::U64TestField;
-    use crate::polynomial::Polynomial;
     use crate::{fft::bit_reversing::in_place_bit_reverse_permute, field::traits::RootsConfig};
     use proptest::{collection, prelude::*};
 
@@ -103,28 +103,6 @@ mod tests {
 
     type F = U64TestField;
     type FE = FieldElement<F>;
-
-    /// Calculates the (non-unitary) Discrete Fourier Transform of `input` via the DFT matrix.
-    fn dft<F: IsTwoAdicField>(input: &[FieldElement<F>]) -> Vec<FieldElement<F>> {
-        let n = input.len();
-        let order = log2(n).unwrap();
-
-        let twiddles = F::get_powers_of_primitive_root(order, n, RootsConfig::Natural).unwrap();
-
-        let mut output = Vec::with_capacity(n);
-        for row in 0..n {
-            let mut sum = FieldElement::zero();
-
-            for (col, element) in input.iter().enumerate() {
-                let i = (row * col) % n; // w^i = w^(i mod n)
-                sum += element.clone() * twiddles[i].clone();
-            }
-
-            output.push(sum);
-        }
-
-        output
-    }
 
     prop_compose! {
         fn powers_of_two(max_exp: u8)(exp in 1..max_exp) -> usize { 1 << exp }
@@ -146,7 +124,7 @@ mod tests {
         // Property-based test that ensures NR Radix-2 FFT gives the same result as a naive DFT.
         #[test]
         fn test_nr_2radix_fft_matches_naive_eval(coeffs in field_vec(8)) {
-            let expected = dft(&coeffs);
+            let expected = naive_matrix_dft_test(&coeffs);
 
             let order = log2(coeffs.len()).unwrap();
             let twiddles = F::get_twiddles(order, RootsConfig::BitReverse).unwrap();
@@ -163,7 +141,7 @@ mod tests {
         // Property-based test that ensures RN Radix-2 FFT gives the same result as a naive DFT.
         #[test]
         fn test_rn_2radix_fft_matches_naive_eval(coeffs in field_vec(8)) {
-            let expected = dft(&coeffs);
+            let expected = naive_matrix_dft_test(&coeffs);
 
             let order = log2(coeffs.len()).unwrap();
             let twiddles = F::get_twiddles(order, RootsConfig::Natural).unwrap();
@@ -173,21 +151,6 @@ mod tests {
             in_place_rn_2radix_fft(&mut result, &twiddles);
 
             prop_assert_eq!(result, expected);
-        }
-    }
-
-    proptest! {
-        // Property-based test that ensures dft() gives the same result as a naive polynomial evaluation.
-        #[test]
-        fn test_dft_same_as_eval(coeffs in field_vec(8)) {
-            let dft = dft(&coeffs);
-
-            let poly = Polynomial::new(&coeffs);
-            let order = log2(coeffs.len()).unwrap();
-            let twiddles = F::get_powers_of_primitive_root(order, coeffs.len(), RootsConfig::Natural).unwrap();
-            let evals: Vec<FE> = twiddles.iter().map(|x| poly.evaluate(x)).collect();
-
-            prop_assert_eq!(evals, dft);
         }
     }
 }
