+++ math/src/fft/mod.rs
@@ -3,3 +3,6 @@ pub mod bit_reversing;
 pub mod errors;
 pub mod fft_iterative;
 pub(crate) mod helpers;
+
+#[cfg(test)]
+pub(crate) mod test_helpers;
